# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Harde Business School API Docs",
        "description": "Harde Business School Application Programme Interface",
        "termsOfService": "THe api is subject to terms of services and maybe be changed without service notice",
        "contact": {
            "name": "Ekenekiso Leonard Ugbanawaji",
            "url": "https://harde.io/support",
            "email": "leonard@hardeverse.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://hardeuniversity.org/assets/img/logo.png"
        }
    },
    "servers": [
        {
            "url": "http://article.test/api/v1",
            "description": "Article API Local Server. This server is attached to the local environment."
        },
        {
            "url": "http://article.test/api/v1"
        }
    ],
    "paths": {
...
        "/articles/{articleId}/comment": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "The Comment paginated resource  collection",
                "description": "The Comment resource collection",
                "operationId": "Api/CommentController::index",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "The article resource Id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "The resource per page",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The paginated Comment resource collection",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/articles/{articleId}/comment/{commentId}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "The Comment resource",
                "description": "The Comment resource",
                "operationId": "Api/CommentController::show",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "The Article resource ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "The comment resource Id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Comment resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CommentResource": {
                "properties": {
                    "subject": {
                        "description": "The Tag resource subject.",
                        "type": "string"
                    },
                    "url": {
                        "description": "The Tag resource body.",
                        "type": "string"
                    },
                    "author": {
                        "description": "The resource Author.",
                        "type": "string"
                    },
                    "articleId": {
                        "description": "The resource Article reference Id.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "400": {
                "description": "Bad Request.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Bad Request.",
                            "errors": []
                        }
                    }
                }
            },
...
            "403": {
                "description": "Forbidden.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Forbidden.",
                            "errors": []
                        }
                    }
                }
            },
            "404": {
                "description": "Resource Not Found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Resource Not Found.",
                            "errors": []
                        }
                    }
                }
            },
...
            "422": {
                "description": "Unprocessable Entity.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Unprocessable Entity.",
                            "errors": []
                        }
                    }
                }
            },
...
            "500": {
                "description": "Server Error.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Server Error.",
                            "errors": []
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/api/v1/login",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}